works all

+++ include/RiceBitVector.h
@@ -188,12 +188,12 @@ template <sux::util::AllocType AT = sux::util::AllocType::MALLOC> class RiceBitV
                     return result;
                 }
 
-                __uint128_t readFixed(const int width) {
+                uint64_t readFixed(const int width) {
                     // In the worst case, curr_fixed_offset is 7.
                     // Loading only 64 bit here would then fail for width>(64-7)=57
                     __uint128_t fixed;
                     memcpy(&fixed, (uint8_t *)&data + curr_fixed_offset / 8, 16);
-                    __uint128_t res =  (fixed >> curr_fixed_offset % 8) & ((__uint128_t(1) << width) - 1);
+                    uint64_t res =  (fixed >> curr_fixed_offset % 8) & ((__uint128_t(1) << width) - 1);
                     curr_fixed_offset += width;
                     return res;
                 }
